@@ -5,8 +5,8 @@ import java.util.Base64;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
-import javax.el.MethodNotFoundException;
-import javax.validation.Valid;
+import jakarta.el.MethodNotFoundException;
+import jakarta.validation.Valid;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
