@@ -3,8 +3,8 @@ package epermit.commons;
 import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.List;
-import javax.persistence.AttributeConverter;
-import javax.persistence.Converter;
+import jakarta.persistence.AttributeConverter;
+import jakarta.persistence.Converter;
 import com.google.gson.reflect.TypeToken;
 
 @Converter
