@@ -3,7 +3,7 @@ package epermit.ledgerevents.quotacreated;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import javax.persistence.criteria.Predicate;
+import jakarta.persistence.criteria.Predicate;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
