@@ -2,7 +2,7 @@ package epermit.services;
 
 import java.util.ArrayList;
 import java.util.List;
-import javax.persistence.criteria.Predicate;
+import jakarta.persistence.criteria.Predicate;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
