@@ -6,10 +6,10 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import java.util.Set;
-import javax.validation.ConstraintViolation;
-import javax.validation.Validation;
-import javax.validation.Validator;
-import javax.validation.ValidatorFactory;
+import jakarta.validation.ConstraintViolation;
+import jakarta.validation.Validation;
+import jakarta.validation.Validator;
+import jakarta.validation.ValidatorFactory;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -60,7 +60,7 @@ public class PermitCreatedLedgerEventHandlerTest {
         event.setPermitId("UZ-TR-2021-1-1");
         event.setPermitIssuedFor("TR");
         event.setPermitIssuer("UZ");
-        event.setPermitType(PermitType.BILITERAL);
+        event.setPermitType(PermitType.BILATERAL);
         event.setPermitYear(2021);
         event.setPlateNumber("A");
         event.setSerialNumber(1);
@@ -119,7 +119,7 @@ public class PermitCreatedLedgerEventHandlerTest {
         event.setIssuedAt("A");
         event.setCompanyName("A");
         event.setPermitId("UZ-TR-2021-1-1");
-        event.setPermitType(PermitType.BILITERAL);
+        event.setPermitType(PermitType.BILATERAL);
         event.setPermitYear(2021);
         event.setPlateNumber("A");
         event.setSerialNumber(1);
@@ -131,7 +131,7 @@ public class PermitCreatedLedgerEventHandlerTest {
         LedgerPermit p = captor.getValue();
         assertEquals("A", p.getExpireAt());
         assertEquals("UZ-TR-2021-1-1", p.getPermitId());
-        assertEquals(PermitType.BILITERAL, p.getPermitType());
+        assertEquals(PermitType.BILATERAL, p.getPermitType());
         assertEquals("A", p.getCompanyName());
         assertEquals("A", p.getIssuedAt());
         assertEquals("UZ", p.getIssuer());
