@@ -109,7 +109,7 @@ public class PermitControllerIT {
         quota.setActive(true);
         quota.setEndNumber(30);
         quota.setStartNumber(1);
-        quota.setPermitType(PermitType.BILITERAL);
+        quota.setPermitType(PermitType.BILATERAL);
         quota.setPermitYear(2021);
         quota.setPermitIssuer("TR");
         quota.setPermitIssuedFor("UZ");
@@ -119,7 +119,7 @@ public class PermitControllerIT {
             SerialNumber serialNumber = new SerialNumber();
             serialNumber.setSerialNumber(i);
             serialNumber.setAuthorityCode("UZ");
-            serialNumber.setPermitType(PermitType.BILITERAL);
+            serialNumber.setPermitType(PermitType.BILATERAL);
             serialNumber.setPermitYear(2021);
             serialNumber.setState(SerialNumberState.CREATED);
             serialNumbers.add(serialNumber);
@@ -158,7 +158,7 @@ public class PermitControllerIT {
             permit.setCompanyName("ABC");
             permit.setCompanyId("1");
             permit.setIssuedFor("UZ");
-            permit.setPermitType(PermitType.BILITERAL);
+            permit.setPermitType(PermitType.BILATERAL);
             permit.setPermitYear(2021);
             permit.setPlateNumber("06AA1234");
             permit.setExpireAt("31/01/2022");
@@ -193,7 +193,7 @@ public class PermitControllerIT {
         permit.setCompanyId("1");
         permit.setIssuer("TR");
         permit.setIssuedFor("UZ");
-        permit.setPermitType(PermitType.BILITERAL);
+        permit.setPermitType(PermitType.BILATERAL);
         permit.setPermitYear(2021);
         permit.setPlateNumber("06AA1234");
         permit.setExpireAt("31/01/2022");
@@ -213,7 +213,7 @@ public class PermitControllerIT {
         input.setCompanyName("ABC");
         input.setCompanyId("123");
         input.setIssuedFor("UZ");
-        input.setPermitType(PermitType.BILITERAL);
+        input.setPermitType(PermitType.BILATERAL);
         input.setPermitYear(2021);
         input.setPlateNumber("06AA1234");
         ResponseEntity<String> r =
@@ -229,7 +229,7 @@ public class PermitControllerIT {
         permit.setCompanyName("ABC");
         permit.setIssuer("TR");
         permit.setIssuedFor("UZ");
-        permit.setPermitType(PermitType.BILITERAL);
+        permit.setPermitType(PermitType.BILATERAL);
         permit.setPermitYear(2021);
         permit.setPlateNumber("06AA1234");
         permit.setExpireAt("31/01/2022");
@@ -279,7 +279,7 @@ public class PermitControllerIT {
         permit.setCompanyName("ABC");
         permit.setIssuer("UZ");
         permit.setIssuedFor("TR");
-        permit.setPermitType(PermitType.BILITERAL);
+        permit.setPermitType(PermitType.BILATERAL);
         permit.setPermitYear(2021);
         permit.setPlateNumber("06AA1234");
         permit.setExpireAt("31/01/2022");
