@@ -23,7 +23,7 @@ public class RestExceptionHandler  {
 
     @ExceptionHandler(ResponseStatusException.class)
     public ResponseEntity<String> handleResponseStatusError(final ResponseStatusException ex) {
-        return new ResponseEntity<>(ex.getMessage(), ex.getStatus());
+        return new ResponseEntity<>(ex.getMessage(), ex.getStatusCode());
     }
 
     @ResponseStatus(HttpStatus.BAD_REQUEST)
@@ -54,3 +54,4 @@ public class RestExceptionHandler  {
         MDC.remove("epermitErrorId");
     }
 }
+
