@@ -123,7 +123,7 @@ public class PermitServiceTest {
     @Test
     void createPermitTest() {
         CreatePermitInput input = new CreatePermitInput();
-        input.setPermitType(PermitType.BILITERAL);
+        input.setPermitType(PermitType.BILATERAL);
         input.setPermitYear(2021);
         input.setIssuedFor("UZ");
         input.setCompanyId("companyId");
@@ -149,7 +149,7 @@ public class PermitServiceTest {
         assertEquals(LedgerEventType.PERMIT_CREATED, event.getEventType());
         assertEquals("TR", event.getPermitIssuer());
         assertEquals("UZ", event.getPermitIssuedFor());
-        assertEquals(PermitType.BILITERAL, event.getPermitType());
+        assertEquals(PermitType.BILATERAL, event.getPermitType());
         assertEquals(2021, event.getPermitYear());
         assertEquals("123", event.getPreviousEventId());
     }
