@@ -82,7 +82,7 @@ public class PermitServiceIT {
         quota.setEndNumber(100);
         quota.setPermitIssuedFor("FR");
         quota.setPermitIssuer("TR");
-        quota.setPermitType(PermitType.BILITERAL);
+        quota.setPermitType(PermitType.BILATERAL);
         quota.setPermitYear(2021);
         quota.setStartNumber(1);
         ledgerQuotaRepository.save(quota);
@@ -91,7 +91,7 @@ public class PermitServiceIT {
             SerialNumber serialNumber = new SerialNumber();
             serialNumber.setSerialNumber(i);
             serialNumber.setAuthorityCode("FR");
-            serialNumber.setPermitType(PermitType.BILITERAL);
+            serialNumber.setPermitType(PermitType.BILATERAL);
             serialNumber.setPermitYear(2021);
             serialNumber.setState(SerialNumberState.CREATED);
             serialNumbers.add(serialNumber);
@@ -105,7 +105,7 @@ public class PermitServiceIT {
         input.setCompanyId("ABC");
         input.setCompanyName("ABC");
         input.setIssuedFor("FR");
-        input.setPermitType(PermitType.BILITERAL);
+        input.setPermitType(PermitType.BILATERAL);
         input.setPermitYear(2021);
         input.setPlateNumber("ABC");
         permitService.createPermit(input);
@@ -117,7 +117,7 @@ public class PermitServiceIT {
         input.setCompanyId("ABC");
         input.setCompanyName("ABC");
         input.setIssuedFor("FR");
-        input.setPermitType(PermitType.BILITERAL);
+        input.setPermitType(PermitType.BILATERAL);
         input.setPermitYear(2021);
         input.setPlateNumber("ABC");
         CreatePermitResult r = permitService.createPermit(input);
